# Optimize UsedSymbolExtractor performance