Before

```
vendor/bin/composer-dependency-analyser --config composer-dependency-analyser.php
  Time (mean ± σ):      2.218 s ±  0.012 s    [User: 2.103 s, System: 0.114 s]
  Range (min … max):    2.202 s …  2.235 s    10 runs
```

After

```
vendor/bin/composer-dependency-analyser --config composer-dependency-analyser.php
  Time (mean ± σ):      1.386 s ±  0.006 s    [User: 1.287 s, System: 0.097 s]
  Range (min … max):    1.375 s …  1.394 s    10 runs
```