@janedbal 2024-02-19
I believe the time in readme deserves to be lowered here :timer_clock: