@janedbal 2024-02-19
The improvement is even better when tested on phpstan-src:

| 4632 files            | Before     | After      | Improvement   |
|-------------|------------|------------|---------------|
| Time (s)    | 1.479      | 0.569      | 61.5%         |